services/ai-broker/pyproject.toml
[project]
name = "ai-broker"
version = "0.1.0"
description = "FastAPI service powering CollabCanvas.ai /cluster and /outline endpoints"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.33",
  "pydantic-settings>=2.6",
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.9"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uvicorn]
reload = true
factory = false
host = "0.0.0.0"
port = 8000
app = "app.main:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
